+++ quantify_scheduler/backends/qblox/compiler_container.py
@@ -7,16 +7,22 @@ from __future__ import annotations
 from typing import TYPE_CHECKING, Any
 
 from quantify_scheduler.backends.qblox import constants
+from quantify_scheduler.backends.qblox.helpers import _generate_legacy_hardware_config
 from quantify_scheduler.backends.qblox.instrument_compilers import (
     ClusterCompiler,
     LocalOscillatorCompiler,
 )
+from quantify_scheduler.helpers.schedule import _extract_port_clocks_used
 from quantify_scheduler.operations.control_flow_library import ControlFlowOperation
 from quantify_scheduler.schedules.schedule import ScheduleBase
 
 if TYPE_CHECKING:
     from quantify_scheduler import Schedule
     from quantify_scheduler.backends.qblox.compiler_abc import InstrumentCompiler
+    from quantify_scheduler.backends.qblox_backend import QbloxHardwareCompilationConfig
+    from quantify_scheduler.backends.types.qblox import (
+        _LocalOscillatorCompilerConfig,
+    )
     from quantify_scheduler.operations.operation import Operation
     from quantify_scheduler.resources import Resource
 
@@ -113,6 +119,7 @@ class CompilerContainer:
         self,
         name: str,
         instrument_cfg: dict[str, Any],
+        portclock_to_path: dict[str, str],
         latency_corrections: dict[str, float] | None = None,
         distortion_corrections: dict[int, Any] | None = None,
     ) -> None:
@@ -121,11 +128,14 @@ class CompilerContainer:
             name=name,
             total_play_time=self.total_play_time,
             instrument_cfg=instrument_cfg,
+            portclock_to_path=portclock_to_path,
             latency_corrections=latency_corrections,
             distortion_corrections=distortion_corrections,
         )
 
-    def _add_local_oscillator(self, name: str, instrument_cfg: dict[str, Any]) -> None:
+    def _add_local_oscillator(
+        self, name: str, instrument_cfg: _LocalOscillatorCompilerConfig
+    ) -> None:
         self.local_oscillators[name] = LocalOscillatorCompiler(
             parent=self,
             name=name,
@@ -135,7 +145,7 @@ class CompilerContainer:
 
     @classmethod
     def from_hardware_cfg(
-        cls, schedule: Schedule, hardware_cfg: dict
+        cls, schedule: Schedule, hardware_cfg: QbloxHardwareCompilationConfig
     ) -> CompilerContainer:
         """
         Factory method for the CompilerContainer. This is the preferred way to use the
@@ -146,20 +156,27 @@ class CompilerContainer:
         schedule
             The schedule to pass to the constructor.
         hardware_cfg
-            The hardware config.
+            The hardware compilation config.
         """
-        distortion_corrections = hardware_cfg.get("distortion_corrections", {})
-        latency_corrections = hardware_cfg.get("latency_corrections", {})
+        # Extract the old-style hardware config from the CompilationConfig
+        hardware_cfg_legacy = _generate_legacy_hardware_config(
+            schedule=schedule, hardware_cfg=hardware_cfg
+        )
+
+        distortion_corrections = hardware_cfg_legacy.get("distortion_corrections", {})
+        latency_corrections = hardware_cfg_legacy.get("latency_corrections", {})
 
         composite = cls(schedule)
-        for instrument_name, instrument_cfg in hardware_cfg.items():
-            if (
-                not isinstance(instrument_cfg, dict)
-                or "instrument_type" not in instrument_cfg
-            ):
-                continue
+        compiler_configs = hardware_cfg._extract_instrument_compiler_configs(
+            _extract_port_clocks_used(schedule)
+        )
 
-            instrument_type = instrument_cfg["instrument_type"]
+        for (
+            instrument_name,
+            instrument_cfg,
+        ) in compiler_configs.items():
+
+            instrument_type = instrument_cfg.instrument_type
 
             instrument_distortion_corrections = None
 
@@ -169,13 +186,15 @@ class CompilerContainer:
             if instrument_type == "Cluster":
                 composite._add_cluster(
                     name=instrument_name,
-                    instrument_cfg=instrument_cfg,
+                    instrument_cfg=hardware_cfg_legacy[instrument_name],
+                    portclock_to_path=instrument_cfg.portclock_to_path,
                     latency_corrections=latency_corrections,
                     distortion_corrections=instrument_distortion_corrections,
                 )
             elif instrument_type == "LocalOscillator":
                 composite._add_local_oscillator(
-                    name=instrument_name, instrument_cfg=instrument_cfg
+                    name=instrument_name,
+                    instrument_cfg=instrument_cfg,
                 )
             else:
                 raise ValueError(