+++ quantify_scheduler/backends/types/qblox.py
@@ -236,48 +236,6 @@ class LOSettings(DataClassJsonMixin):
     frequency: Dict[str, Optional[float]]
     """The frequency to set the LO to."""
 
-    @classmethod
-    def from_mapping(cls, mapping: Dict[str, Any]) -> LOSettings:
-        """
-        Factory method for the LOSettings from a mapping dict. The required format is
-        {"frequency": {parameter_name: value}, "power": {parameter_name: value}}. For
-        convenience {"frequency": value, "power": value} is also allowed.
-
-        Parameters
-        ----------
-        mapping
-            The part of the mapping dict relevant for this instrument.
-
-        Returns
-        -------
-        :
-            Instantiated LOSettings from the mapping dict.
-        """
-        if "power" not in mapping:
-            raise KeyError(
-                "Attempting to compile settings for a local oscillator but 'power' is "
-                "missing from settings. 'power' is required as an entry for Local "
-                "Oscillators."
-            )
-        if "generic_icc_name" in mapping:
-            generic_icc_name = mapping["generic_icc_name"]
-            if generic_icc_name != constants.GENERIC_IC_COMPONENT_NAME:
-                raise NotImplementedError(
-                    f"Specified name '{generic_icc_name}' as a generic instrument "
-                    f"coordinator component, but the Qblox backend currently only "
-                    f"supports using the default name "
-                    f"'{constants.GENERIC_IC_COMPONENT_NAME}'"
-                )
-
-        power_entry: Union[float, Dict[str, float]] = mapping["power"]
-        if not isinstance(power_entry, dict):  # floats allowed for convenience
-            power_entry = {"power": power_entry}
-        freq_entry: float | None | Dict[str, float | None] = mapping.get("frequency")
-        if not isinstance(freq_entry, dict):
-            freq_entry = {"frequency": freq_entry}
-
-        return cls(power=power_entry, frequency=freq_entry)
-
 
 _ModuleSettingsT = TypeVar("_ModuleSettingsT", bound="BaseModuleSettings")
 """